backend/requirements.txt
fastapi
uvicorn[standard]
dbt-bigquery
dotenv
google-cloud-bigquery
google-cloud-storage
google-cloud-monitoring
scikit-learn
pandas
numpy
lightgbm
pydantic>=2.0.0
pydantic-settings>=2.0.0